Update all requests to use cakePHP http client

+++ src/Feature/OrangeApi.php
@@ -14,14 +14,13 @@ abstract class OrangeApi
 
     public function __construct(Authorization $authorization, string $logPath = null)
     {
-        if ($authorization === null) {
-            throw new \RuntimeException('authorization to orange api host must be provided.');
-        }
-
         $this->authorization = $authorization;
 
         if ($logPath !== null) {
-            $stream = new StreamHandler($logPath . '/sms_reporting_'. gmdate('Ymd') . '.log', Logger::DEBUG);
+            $stream = new StreamHandler(
+                $logPath . '/sms_reporting_'. gmdate('Ymd') . '.log',
+                Logger::DEBUG
+            );
             $firephp = new FirePHPHandler();
 
             $this->logger = new Logger($this->authorization->getClientId());
@@ -47,12 +46,8 @@ abstract class OrangeApi
     {
         $callResponse = $this->query($args);
 
-        if (!array_key_exists('code', $callResponse) || !array_key_exists('response', $callResponse)) {
-            throw new \RuntimeException('Response is malformed.');
-        }
-
-        if ($callResponse['code'] !== $response_code) {
-            if ($callResponse['code'] === 401 && $callResponse['response']['code'] === 42) {
+        if (array_key_exists('code', $callResponse) && $callResponse['code'] !== $response_code) {
+            if ($callResponse['code'] === 42) {
                 unlink($this->authorization->getLogPath());
 
                 if ($this->isAuthorized()) {
@@ -60,6 +55,7 @@ abstract class OrangeApi
                 }
             }
         }
+
         return $callResponse;
     }
 }
